+++ apps/ensadmin/src/components/indexing-status/indexing-status.tsx
@@ -7,10 +7,10 @@
 import { OverallIndexingStatusIds } from "@ensnode/ensnode-sdk";
 import { type ReactElement, Suspense } from "react";
 
-import { useENSIndexerConfig, useIndexingStatus } from "@/components/ensindexer/hooks";
 import { RecentRegistrations } from "@/components/recent-registrations";
 import { Card, CardContent, CardHeader } from "@/components/ui/card";
 
+import { useENSIndexerConfig, useIndexingStatus } from "@ensnode/ensnode-react";
 import { BackfillStatus } from "./backfill-status";
 import { ENSIndexerDependencyInfo } from "./dependecy-info";
 import {
