+++ packages/ensnode-react/src/hooks/index.ts
@@ -2,3 +2,5 @@ export * from "./useENSNodeConfig";
 export * from "./useRecords";
 export * from "./usePrimaryName";
 export * from "./usePrimaryNames";
+export * from "./useENSIndexerConfig";
+export * from "./useIndexingStatus";
